requirements.txt
django~=5.1.2
djangorestframework~=3.15.2
PyMySQL~=1.1.1
streamlit~=1.39.0
requests~=2.32.3